update to langcodes library

+++ source/add_srt_to_video/plugin.py
@@ -24,6 +24,8 @@ import os
 import glob
 import difflib
 import subprocess
+from langcodes.tag_parser import LanguageTagError
+from langcodes import *
 
 from unmanic.libs.unplugins.settings import PluginSettings
 
@@ -32,24 +34,29 @@ from add_srt_to_video.lib.ffmpeg import Probe, Parser
 # Configure plugin logger
 logger = logging.getLogger("Unmanic.Plugin.srt_to_video")
 
-lang_codes = [('aa', 'aar'), ('ab', 'abk'), ('af', 'afr'), ('ak', 'aka'), ('am', 'amh'), ('ar', 'ara'), ('an', 'arg'), ('as', 'asm'), ('av', 'ava'), ('ae', 'ave'), ('ay', 'aym'), ('az', 'aze'), ('ba', 'bak'), ('bm', 'bam'), ('be', 'bel'), ('bn', 'ben'), ('bi', 'bis'), ('bo', 'bod / tib*'),
-              ('bs', 'bos'), ('br', 'bre'), ('bg', 'bul'), ('ca', 'cat'), ('cs', 'ces / cze*'), ('ch', 'cha'), ('ce', 'che'), ('cn', 'zho / chi*'), ('cu', 'chu'), ('cv', 'chv'), ('kw', 'cor'), ('co', 'cos'), ('cr', 'cre'), ('cy', 'cym / wel*'), ('da', 'dan'), ('de', 'deu / ger*'), ('dv', 'div'), ('dz', 'dzo'),
-              ('el', 'ell / gre*'), ('en', 'eng'), ('eo', 'epo'), ('et', 'est'), ('eu', 'eus / baq*'), ('ee', 'ewe'), ('fo', 'fao'), ('fa', 'fas / per*'), ('fj', 'fij'), ('fi', 'fin'), ('fr', 'fra / fre*'), ('fy', 'fry'), ('ff', 'ful'), ('gd', 'gla'), ('ga', 'gle'), ('gl', 'glg'), ('gv', 'glv'),
-              ('gn', 'grn'), ('gu', 'guj'), ('ht', 'hat'), ('ha', 'hau'), ('he', 'heb'), ('hz', 'her'), ('hi', 'hin'), ('ho', 'hmo'), ('hr', 'hrv'), ('hu', 'hun'), ('hy', 'hye / arm*'), ('ig', 'ibo'), ('io', 'ido'), ('ii', 'iii'), ('iu', 'iku'), ('ie', 'ile'), ('ia', 'ina'), ('id', 'ind'),
-              ('ik', 'ipk'), ('is', 'isl / ice*'), ('it', 'ita'), ('jv', 'jav'), ('ja', 'jpn'), ('kl', 'kal'), ('kn', 'kan'), ('ks', 'kas'), ('ka', 'kat / geo*'), ('kr', 'kau'), ('kk', 'kaz'), ('km', 'khm'), ('ki', 'kik'), ('rw', 'kin'), ('ky', 'kir'), ('kv', 'kom'), ('kg', 'kon'), ('ko', 'kor'),
-              ('kj', 'kua'), ('ku', 'kur'), ('lo', 'lao'), ('la', 'lat'), ('lv', 'lav'), ('li', 'lim'), ('ln', 'lin'), ('lt', 'lit'), ('lb', 'ltz'), ('lu', 'lub'), ('lg', 'lug'), ('mh', 'mah'), ('ml', 'mal'), ('mr', 'mar'), ('mk', 'mkd / mac*'), ('mg', 'mlg'), ('mt', 'mlt'), ('mn', 'mon'),
-              ('mi', 'mri / mao*'), ('ms', 'msa / may*'), ('my', 'mya / bur*'), ('na', 'nau'), ('nv', 'nav'), ('nr', 'nbl'), ('nd', 'nde'), ('ng', 'ndo'), ('ne', 'nep'), ('nl', 'nld / dut*'), ('nn', 'nno'), ('nb', 'nob'), ('no', 'nor* / nob / nno'), ('ny', 'nya'), ('oc', 'oci'), ('oj', 'oji'), ('or', 'ori'),
-              ('om', 'orm'), ('os', 'oss'), ('pa', 'pan'), ('pi', 'pli'), ('pl', 'pol'), ('pt', 'por'), ('ps', 'pus'), ('qu', 'que'), ('rm', 'roh'), ('ro', 'ron / rum*'), ('rn', 'run'), ('ru', 'rus'), ('sg', 'sag'), ('sa', 'san'), ('si', 'sin'), ('sk', 'slk / slo*'), ('sl', 'slv'), ('se', 'sme'),
-              ('sm', 'smo'), ('sn', 'sna'), ('sd', 'snd'), ('so', 'som'), ('st', 'sot'), ('es', 'spa'), ('sq', 'sqi / alb*'), ('sc', 'srd'), ('sr', 'srp'), ('ss', 'ssw'), ('su', 'sun'), ('sw', 'swa'), ('sv', 'swe'), ('ty', 'tah'), ('ta', 'tam'), ('tt', 'tat'), ('te', 'tel'), ('tg', 'tgk'),
-              ('tl', 'tgl'), ('th', 'tha'), ('ti', 'tir'), ('to', 'ton'), ('tn', 'tsn'), ('ts', 'tso'), ('tk', 'tuk'), ('tr', 'tur'), ('tw', 'twi'), ('ug', 'uig'), ('uk', 'ukr'), ('ur', 'urd'), ('uz', 'uzb'), ('ve', 'ven'), ('vi', 'vie'), ('vo', 'vol'), ('wa', 'wln'), ('wo', 'wol'), ('xh', 'xho'),
-              ('yi', 'yid'), ('yo', 'yor'), ('za', 'zha'), ('zh', 'zho / chi*'), ('zu', 'zul')]
-
 class Settings(PluginSettings):
     settings = {
+        "tag_style": "3",
     }
 
     def __init__(self, *args, **kwargs):
         super(Settings, self).__init__(*args, **kwargs)
+        self.form_settings = {
+            "tag_style": {
+                "label":          "Choose Language Tag Style",
+                "description":    "Select whether you prefer language tags to be written as either '2 letter' or '3 letter' codes",
+                "input_type":     "select",
+                "select_options": [
+                    {
+                        "value": "2",
+                        "label": "2 letter",
+                    },
+                    {   "value": "3",
+                        "label": "3 letter",
+                    }
+                ],
+            }
+        }
 
 def on_library_management_file_test(data):
     """
@@ -84,14 +91,6 @@ def on_library_management_file_test(data):
             data['add_file_to_pending_tasks'] = True
         return data
 
-def lang_split(lang):
-    l = lang.split('/')
-    l = [i.strip() for i in l]
-    if len(l) > 1:
-        l = [i for i in l if '*' in i]
-    l = [l2.replace('*','') for l2 in l]
-    return l[0]
-
 def check_sub(subfile, encoder, suffix):
     if suffix == '.mkv':
         fmt = 'matroska'
@@ -123,6 +122,14 @@ def on_worker_process(data):
     # Default to no FFMPEG command required. This prevents the FFMPEG command from running if it is not required
     data['exec_command'] = []
     data['repeat'] = False
+
+    # Configure settings object (maintain compatibility with v1 plugins)
+    if data.get('library_id'):
+        settings = Settings(library_id=data.get('library_id'))
+    else:
+        settings = Settings()
+
+    tag_style = settings.get_setting('tag_style')
 
     # Get the path to the input and output files
     abspath = data.get('file_in')
@@ -164,45 +171,44 @@ def on_worker_process(data):
         if srt_to_skip:
             srt_files = [srt_files[i] for i in range(len(srt_files)) if i not in srt_to_skip]
 
-        # get all subtitle files in folder where original video file is, get 3 letter language code, build ffmpeg subtitle args for new streams
+        # get all subtitle files in folder where original video file is, get 2 or 3 letter language code as specified by config, build ffmpeg subtitle args for new streams
         for j in range(len(srt_files)):
-            ffmpeg_args += ['-i', str(srt_files[j])]
             lang_srt = [li for li in difflib.ndiff(basefile, srt_files[j]) if li[0] != ' ']
             # lang = ''.join([i.replace('+ ','') for i in lang_srt]).replace('.srt','').replace('.','')
             langbase = ''.join([i.replace('+ ','') for i in lang_srt])
-            lang = langbase.split('.')[1]
-            if len(lang) == 2:
-                try:
-                    lang3 = [lang_codes[i][1] for i in range(len(lang_codes)) if lang == lang_codes[i][0]][0]
-                    lang3 = lang_split(lang3)
-                except (KeyError, IndexError):
-                    logger.error("error translating language code of subtitle stream - aborting. language code: '{}'".format(lang))
-                    return data
-            elif len(lang) == 3:
-                try:
-                    lang3 = [lang_codes[i][1] for i in range(len(lang_codes)) if lang in lang_codes[i][1]][0]
-                    lang3 = lang_split(lang3)
-                except (KeyError, IndexError):
-                    logger.error("error - 3 letter language code provided not found - aborting. language code: '{}'".format(lang))
-                    return data
+            langbase_tag = langbase.split('.')[1]
+
+            try:
+                lang_tag = standardize_tag(langbase_tag) if Language.get(langbase_tag).is_valid() else ""
+            except LanguageTagError:
+                lang_tag = ""
+
+            logger.debug(f"lang_tag: {lang_tag}, j: {j}, langbase: {langbase}")
+
+            if lang_tag:
+                if tag_style == "2":
+                    lang_tag = standardize_tag(lang_tag)
+                else:
+                    lang_tag = Language.get(standardize_tag(lang_tag)).to_alpha3()
+                ffmpeg_args += ['-i', str(srt_files[j])]
+                ffmpeg_subtitle_args += ['-map', '{}:s:0'.format(j+1), '-c:s:{}'.format(j+existing_subtitle_streams_list_len), str(encoder), '-metadata:s:s:{}'.format(j+existing_subtitle_streams_list_len), 'language={}'.format(lang_tag)]
             else:
-                logger.error("error - improper langauge code supplied - aborting. language code: '{}'".format(lang))
-                return data
-            ffmpeg_subtitle_args += ['-map', '{}:s:0'.format(j+1), '-c:s:{}'.format(j+existing_subtitle_streams_list_len), str(encoder), '-metadata:s:s:{}'.format(j+existing_subtitle_streams_list_len), 'language={}'.format(lang3)]
-
-        # add in any existing subtitle streams
-        for i in range(existing_subtitle_streams_list_len-1, -1, -1):
-            ffmpeg_subtitle_args = ['-map', '0:s:{}'.format(i), '-c:s:{}'.format(i), 'copy'] + ffmpeg_subtitle_args
-
-        # build rest of ffmpeg_args around ffmpeg_subtitle_args
-        ffmpeg_args += ['-max_muxing_queue_size', '9999', '-strict', '-2', '-map', '0:v', '-c:v', 'copy', '-map', '0:a', '-c:a', 'copy'] + ffmpeg_subtitle_args + ['-map', '0:t?', '-c:t', 'copy', '-map', '0:d?', '-c:d', 'copy']
-        if sfx == mp4:
-            ffmpeg_args += ['-dn', '-map_metadata:c', '-1', '-y', str(outfile)]
-        else:
-            ffmpeg_args += ['-y', str(outfile)]
+                logger.info(f"Unable to identify language code {langbase_tag} from srt extension {langbase} - skipping stream")
 
         if ffmpeg_subtitle_args:
-            # external subtitle file(s) were found for video file
+        # external subtitle file(s) were found for video file
+
+            # add in any existing subtitle streams
+            for i in range(existing_subtitle_streams_list_len-1, -1, -1):
+                ffmpeg_subtitle_args = ['-map', '0:s:{}'.format(i), '-c:s:{}'.format(i), 'copy'] + ffmpeg_subtitle_args
+
+            # build rest of ffmpeg_args around ffmpeg_subtitle_args
+            ffmpeg_args += ['-max_muxing_queue_size', '9999', '-strict', '-2', '-map', '0:v', '-c:v', 'copy', '-map', '0:a', '-c:a', 'copy'] + ffmpeg_subtitle_args + ['-map', '0:t?', '-c:t', 'copy', '-map', '0:d?', '-c:d', 'copy']
+            if sfx == mp4:
+                ffmpeg_args += ['-dn', '-map_metadata:c', '-1', '-y', str(outfile)]
+            else:
+                ffmpeg_args += ['-y', str(outfile)]
+
             logger.debug("ffmpeg_args: '{}'".format(ffmpeg_args))
 
             # Apply ffmpeg args to command
@@ -241,6 +247,18 @@ def on_postprocessor_task_results(data):
     status = data.get('task_processing_success')
     logger.debug("status: '{}'".format(status))
 
+    destpath = data.get('destination_files')[0]
+    probe = Probe(logger, allowed_mimetypes=['video'])
+    if probe.file(destpath):
+        probe_streams = probe.get_probe()["streams"]
+    else:
+        logger.error("Cannot probe destination file so cannot reliably delete srt files.  Source file: {data.get('source_data').get('abspath')}")
+        return data
+
+    astreams = [s for s in probe_streams if s.get('codec_type') == 'audio']
+    langs = [s.get('tags').get('language') for s in astreams or {}]
+    std_langs = [str(Language.get(standardize_tag(lt))) for lt in langs if Language.get(lt).is_valid() or None]
+
     if status:
         abspath = data.get('source_data').get('abspath')
         basefile = os.path.splitext(abspath)[0]
@@ -249,7 +267,12 @@ def on_postprocessor_task_results(data):
         logger.debug("srt files: '{}'".format(srt_files))
         for j in range(len(srt_files)):
             srt_file = srt_files[j]
-            os.remove(srt_file)
-            logger.info("srt file '{}' has been added to video file of the same basename; the srt file has been deleted.".format(srt_file))
-
+            lang_srt = [li for li in difflib.ndiff(basefile, srt_files[j]) if li[0] != ' ']
+            lang_tag = ''.join([i.replace('+ ','') for i in lang_srt]).replace('.srt','').replace('.','')
+            std_lang = str(Language.get(standardize_tag(lang_tag))) if Language.get(lang_tag).is_valid() else None
+            if std_lang in std_langs:
+                os.remove(srt_file)
+                logger.info("srt file '{}' has been added to video file of the same basename; the srt file has been deleted.".format(srt_file))
+            else:
+                logger.info(f"srt file {srt_file} language was not found in destination file, so not removing")
     return data
